// src/types/user.ts
export type RegUser = {
  id: number;
  name: string;
  email: string;
  role: string;
  avatar_url: string;
  client_id: number;
  created_at: string;
  user_id: string;
  first_name: string;
  last_name: string;
  phone_number: string;
  timezone: string;
  language: string;
  password_updated_at: string;
  mfa_enabled: boolean;
};